requirements.txt
# Core dependencies
torch==2.2.0
transformers==4.36.0
datasets==2.16.0
accelerate==0.25.0
tokenizers==0.15.0

# Training utilities
tensorboard==2.15.1
tqdm==4.66.1
pyyaml==6.0.1

# Numerical computing
numpy>=1.26.0
scipy==1.11.4

# Logging and monitoring
wandb==0.16.1
rich==13.7.0

# Development tools
black==23.12.0
isort==5.13.2
pylint==3.0.3
pytest==7.4.3
pre-commit==3.6.0

# Additional utilities
python-dotenv==1.0.0
omegaconf==2.3.0
black==23.12.0
